Attempts to fix use in VR
-For an unknown reason, IK isn't working properly. Could be due to the
transition to the Vive from the Oculus Touch, but will have to see
later.
-Visualizing drawn splines in VR isn't working, but may be due to
splines not being drawn in the first place. This could be due to an
incorrect use of input onthe Vive, or just a general lack of knowledge
and improper implementation of handling splines and visualizing them.
-Harrison